refactor(ui): remove `useNamespaceManager` composable
